+++ examples/android/tflite_convertor/tfltransfer/model_correctness_test.py
@@ -149,7 +149,7 @@ class TransferModel(object):
         [x_in] = interpreter.get_input_details()
         [bottleneck_out] = interpreter.get_output_details()
 
-        for (x, y) in image_gen:
+        for x, y in image_gen:
             batch_size = x.shape[0]
             interpreter.resize_tensor_input(
                 x_in["index"], (batch_size, IMAGE_SIZE, IMAGE_SIZE, 3)