+++ src/py/flwr_example/pytorch_save_weights/client.py
@@ -30,6 +30,8 @@ from . import cifar
 
 # pylint: disable=no-member
 DEVICE: str = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+
+
 # Flower Client
 class CifarClient(fl.client.NumPyClient):
     """Flower client implementing CIFAR-10 image classification using PyTorch."""